% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DD_to_DMS.R
\name{DD_to_DMS}
\alias{DD_to_DMS}
\title{DD_to_DMS}
\usage{
DD_to_DMS(DD_input, axis)
}
\arguments{
\item{DD_input}{latitude or longitude in decimal degrees (numeric)}

\item{axis}{one of \code{"horizontal"} (latitude) or \code{"vertical"} (longitude)}
}
\value{
DMS
}
\description{
DD_to_DMS
}
\examples{
DD_to_DMS(DD_input = -63, axis = "horizontal")
DD_to_DMS(DD_input = 44, axis = "vertical")
}
