% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DDM_to_DMS.R
\name{DDM_to_DMS}
\alias{DDM_to_DMS}
\title{DDM_to_DMS}
\usage{
DDM_to_DMS(DDM_input, axis)
}
\arguments{
\item{DDM_input}{latitude or longitude in degrees decimal minutes (character); e.g. \code{"63° 0 W"}}

\item{axis}{one of \code{"horizontal"} (latitude) or \code{"vertical"} (longitude)}
}
\value{
DMS
}
\description{
DDM_to_DMS
}
\examples{
DDM_to_DMS(DDM_input = "63° 0 W", axis = "horizontal")
DDM_to_DMS(DDM_input = "44° 0 N", axis = "vertical")
}
