% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DDM_to_UTM.R
\name{DDM_to_UTM}
\alias{DDM_to_UTM}
\title{DDM_to_UTM}
\usage{
DDM_to_UTM(x, y, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{latitude in degrees decimal minutes (character); e.g. \code{"63° 0 W"}}

\item{y}{longitude in degrees decimal minutes (character); e.g. \code{"44° 0 N"}}

\item{ellipsoid}{reference ellipsoid (character); defaults to \code{"WGS40"}}

\item{return}{desired output; one of \code{"all"}, \code{"northing"}, \code{"easting"}, or \code{"zone"}}
}
\value{
UTM
}
\description{
DDM_to_UTM
}
\examples{
DDM_to_UTM(x = "63° 0 W", y = "44° 0 N", return = "all")
DDM_to_UTM(x = "63° 0 W", y = "44° 0 N", return = "northing")
DDM_to_UTM(x = "63° 0 W", y = "44° 0 N", return = "easting")
DDM_to_UTM(x = "63° 0 W", y = "44° 0 N", return = "zone")
}
