% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DMS_to_DD.R
\name{DMS_to_DD}
\alias{DMS_to_DD}
\title{DMS_to_DD}
\usage{
DMS_to_DD(DMS_input, axis)
}
\arguments{
\item{DMS_input}{latitude or longitude in degrees minute second format (character); e.g. \verb{"63° 0' 0\\" E"}}

\item{axis}{one of \code{"horizontal"} (latitude) or \code{"vertical"} (longitude)}
}
\value{
DD
}
\description{
DMS_to_DD
}
\examples{
DMS_to_DD(DMS_input = "63° 0' 0\" E", axis = "horizontal")
DMS_to_DD(DMS_input = "44° 0' 0\" N", axis = "vertical")
}
