% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DMS_to_UTM.R
\name{DMS_to_UTM}
\alias{DMS_to_UTM}
\title{DMS_to_UTM}
\usage{
DMS_to_UTM(x, y, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{latitude in degrees minutes seconds (character); e.g. \verb{"63° 0' 0\\" E"}}

\item{y}{longitude in degrees minutes seconds (character); e.g. \verb{"44° 0' 0\\" W"}}

\item{ellipsoid}{reference ellipsoid (character); defaults to \code{"WGS40"}}

\item{return}{desired output; one of \code{"all"}, \code{"northing"}, \code{"easting"}, or \code{"zone"}}
}
\value{
UTM
}
\description{
DMS_to_UTM
}
\examples{
DMS_to_UTM(x = "63° 0' 0\" E", y = "44° 0' 0\" W", return = "all")
DMS_to_UTM(x = "63° 0' 0\" E", y = "44° 0' 0\" W", return = "northing")
DMS_to_UTM(x = "63° 0' 0\" E", y = "44° 0' 0\" W", return = "easting")
DMS_to_UTM(x = "63° 0' 0\" E", y = "44° 0' 0\" W", return = "zone")
}
