% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_UTM_to_DDM.R
\name{UTM_to_DDM}
\alias{UTM_to_DDM}
\title{UTM_to_DDM}
\usage{
UTM_to_DDM(x, y, zone, ellipsoid = "WGS84", return = "all")
}
\arguments{
\item{x}{easting (numeric)}

\item{y}{northing (numeric)}

\item{zone}{UTM zone (numeric)}

\item{ellipsoid}{reference ellipsoid (character); defaults to \code{"WGS40"}}

\item{return}{desired output; one of \code{"all"}, \code{"northing"}, \code{"easting"}, or \code{"zone"}}
}
\value{
DDM
}
\description{
UTM_to_DDM
}
\examples{
UTM_to_DDM(x = 271542.3, y = 2582924, zone = 16, return = "all")
UTM_to_DDM(x = 271542.3, y = 2582924, zone = 16, return = "lat")
UTM_to_DDM(x = 271542.3, y = 2582924, zone = 16, return = "lon")
}
